code polishes
