+++ Sources/Playbook/Components/Popover/PopoverHandler.swift
@@ -15,29 +15,13 @@ struct PopoverHandler: ViewModifier {
     
     func body(content: Content) -> some View {
         content
-            .overlay(GlobalPopoverView { popoverID = $0 } )
+            .overlay(GlobalPopoverView { popoverID = $0 })
             .onTapGesture {
                 if let id = popoverID {
-                    closeOutside(id)
+                    popoverManager.closeOutside(id)
                 }
             }
             .environmentObject(popoverManager)
-    }
-    
-    private func closeOutside(_ key: Int){
-        switch popoverManager.popovers[key]?.close.0 {
-            case .outside, .anywhere:
-                popoverManager.isPresented.updateValue(false, forKey: key)
-                
-            default:
-                break
-        }
-    }
-}
-
-public extension View {
-    func popoverHandler() -> some View {
-        self.modifier(PopoverHandler())
     }
 }
 
@@ -50,14 +34,14 @@ struct GlobalPopoverView: View {
             let presentedList = popoverManager.isPresented.sorted(by: { $0.key <= $1.key })
             let popoverList = popoverManager.popovers.sorted(by: { $0.key <= $1.key })
             let commonKeys = Set(presentedList.map { $0.key }).intersection(popoverList.map { $0.key })
-            let newArray: [(Int, Bool, PopoverManager.Popover)] = commonKeys.compactMap { key in
+            let compactArray: [(Int, Bool, PopoverManager.Popover)] = commonKeys.compactMap { key in
                 if let presentedValue = popoverManager.isPresented[key], let listValue = popoverManager.popovers[key] {
                     return (key, presentedValue, listValue)
                 }
                 return nil
             }
             
-            ForEach(newArray, id: \.0) { id, isPresented, popover in
+            ForEach(compactArray, id: \.0) { id, isPresented, popover in
                 if isPresented {
                     Color.clear
                         .frame(maxWidth: .infinity, maxHeight: .infinity)
@@ -65,42 +49,16 @@ struct GlobalPopoverView: View {
                     popover.view
                         .position(popover.position ?? .zero)
                         .onTapGesture {
-                            closeInside(id)
+                            popoverManager.closeInside(id)
                         }
                 }
-                
             }
-            
-        }
-        
-    }
-    
-    private func closeInside(_ key: Int) -> Void {
-        switch popoverManager.popovers[key]?.close.0 {
-            case .inside, .anywhere:
-                onClose(key)
-            default:
-                break
         }
     }
-    
-    private func closeOutside(_ key: Int) -> AnyView {
-        switch popoverManager.popovers[key]?.close.0 {
-            case .outside, .anywhere:
-                return AnyView(ClickDetectorView {
-                    onClose(key)
-                })
-            default:
-                return AnyView(EmptyView())
-        }
-    }
-    
-    private func onClose(_ key: Int) {
-        if let closeAction = popoverManager.popovers[key]?.close.action {
-            popoverManager.isPresented.updateValue(false, forKey: key)
-            closeAction()
-        } else {
-            popoverManager.isPresented.updateValue(false, forKey: key)
-        }
+}
+
+public extension View {
+    func popoverHandler() -> some View {
+        self.modifier(PopoverHandler())
     }
 }
